* Improved temp region patch (Fixes one region only games that don't work in other regions.. (The Amazing Race and Hollywood Squares on PAL nands for example)) * Added temp region patch for shop channel (found by JoostinOnline)
Note: The temp region only patch for the shop channel is untested, but I asume it works. Will test tomorrow and if needed change it.

Users have to provide a setting.txt file with a matching serial# of their own Wii. ModMii by XFlak is able to create an emu nand with a setting.txt file with a matching serial#. To enable the patch simply choose your own region in the overlay menu (Menu Hacks) were it says: "Patch Shop Channel For Region:" and neek will take care of the rest

git-svn-id: svn://localhost/Users/maga/Downloads/custom-di/trunk@76 cd6a0119-72aa-4de0-b879-bd60aad56387